requirements-test.txt
Pillow>=10.0.0
pyzbar>=0.1.9
pytest>=8.0.0
